Work out local IP automatically

+++ bin/run_tests.sh
@@ -5,6 +5,7 @@ function in_docker(){
 }
 
 ROOT_DIR=$(realpath $(dirname "${BASH_SOURCE[0]}")/..)
+ONLY_DRIVER=false
 
 function log(){
     local fatal=false
@@ -28,9 +29,26 @@ function log(){
     fi
 }
 
+function get_local_ip(){
+    ifconfig | grep inet | grep -v inet6 | grep -v 127.0.0.1 | awk '{print $2}' | head -n 1
+}
+
 function run_containers(){
     local RESULT
     local STARTED_DRIVER=false
+    local ONLY_DRIVER=false
+
+    while true; do
+        case "$1" in
+            --only-driver)
+                ONLY_DRIVER=true
+                shift
+                ;;
+            *)
+                break
+                ;;
+        esac
+    done
 
     if ! docker ps -a --format "table {{.Names}}\t{{.Image}}\t{{.Status}}" | grep -q selenium-driver; then
         log "Starting driver"
@@ -42,44 +60,45 @@ function run_containers(){
             --shm-size=2g \
             seleniarm/standalone-firefox
 
-        sleep 2
+        RESULT=$?
 
         STARTED_DRIVER=true
+
+        sleep 2
     fi
 
-    log "Deleting old screenshots"
+    if ! $ONLY_DRIVER; then
+        log "Deleting old screenshots"
 
-    rm -fr $ROOT_DIR/screenshots
-    mkdir -p $ROOT_DIR/screenshots
+        rm -fr $ROOT_DIR/screenshots
+        mkdir -p $ROOT_DIR/screenshots
 
-    log "Running tests"
+        log "Running tests"
 
-    docker run -it --rm \
-        -v $ROOT_DIR/sites:/opt/paysauce-tests/sites \
-        -v $ROOT_DIR/screenshots:/opt/paysauce-tests/screenshots \
-        -v $ROOT_DIR/bin:/opt/paysauce-tests/bin \
-        paysauce-tests $*
+        docker run -it --rm \
+            -v $ROOT_DIR/sites:/opt/paysauce-tests/sites \
+            -v $ROOT_DIR/screenshots:/opt/paysauce-tests/screenshots \
+            -v $ROOT_DIR/bin:/opt/paysauce-tests/bin \
+            -e SELENIUM_HUB_HOST=$(get_local_ip) \
+            paysauce-tests $*
 
-    RESULT=$?
+        RESULT=$?
 
-    if $STARTED_DRIVER; then
-        log "Stopping driver"
+        if $STARTED_DRIVER; then
+            log "Stopping driver"
 
-        docker stop selenium-driver
+            docker stop selenium-driver
+        fi
     fi
 
     return $RESULT
 }
 
-if ! in_docker; then
-    run_containers $*
-    exit $?
-fi
-
 while true; do
     case "$1" in
         --help|-h)
-            echo "TODO: helptext"
+            echo "Usage:   ./bin/run_tests.sh [test-dir]"
+            echo "Example: ./bin/run_tests.sh www.paysauce.com"
             exit 0
             ;;
         *)
@@ -87,6 +106,11 @@ while true; do
             ;;
     esac
 done
+
+if ! in_docker; then
+    run_containers $*
+    exit $?
+fi
 
 function run_tests_for(){
     local DIR="$1"
